Mathy/Mathy/Resources/requirements.txt
pix2tex>=0.1.1
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
